+++ source/gatery/hlim/coreNodes/Node_Register.cpp
@@ -151,7 +151,7 @@ void Node_Register::estimateSignalDelay(SignalDelay &sigDelay)
         f = 0.0f;
 }
 
-void Node_Register::estimateSignalDelayCriticalInput(SignalDelay &sigDelay, unsigned outputPort, unsigned outputBit, unsigned &inputPort, unsigned &inputBit)
+void Node_Register::estimateSignalDelayCriticalInput(SignalDelay &sigDelay, size_t outputPort, size_t outputBit, size_t &inputPort, size_t &inputBit)
 {
     inputPort = ~0u;
     inputBit = ~0u;
